third_party/blink/web_tests/svg/dynamic-updates/SVGFETileElement-dom-in-attr-expected.txt

SVG 1.1 dynamic update tests

Tests dynamic updates of the 'in' attribute of the SVGFETileElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS tile.getAttribute('in') is "SourceGraphic"
PASS tile.getAttribute('in') is "offset"
PASS successfullyParsed is true

TEST COMPLETE
